refactor: move existing mcp version to new package
Separate MCP protocol version -specific code into it's own package.
MCP will pick which version to use based on the protocol negotiated during initialization.
Will remove the hard-coded protocol in the next PR during mcpManager implementation.

This PR is a no-op change and doesn't add new feature.
