+++ internal/server/api.go
@@ -17,13 +17,13 @@ package server
 import (
 	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 	"github.com/go-chi/render"
 	"github.com/googleapis/genai-toolbox/internal/tools"
+	"github.com/googleapis/genai-toolbox/internal/util"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/metric"
@@ -199,7 +199,7 @@ func toolInvokeHandler(s *Server, w http.ResponseWriter, r *http.Request) {
 	s.logger.DebugContext(ctx, "tool invocation authorized")
 
 	var data map[string]any
-	if err = decodeJSON(r.Body, &data); err != nil {
+	if err = util.DecodeJSON(r.Body, &data); err != nil {
 		render.Status(r, http.StatusBadRequest)
 		err = fmt.Errorf("request body was invalid JSON: %w", err)
 		s.logger.DebugContext(ctx, err.Error())
@@ -273,14 +273,4 @@ type errResponse struct {
 func (e *errResponse) Render(w http.ResponseWriter, r *http.Request) error {
 	render.Status(r, e.HTTPStatusCode)
 	return nil
-}
-
-// decodeJSON decodes a given reader into an interface using the json decoder.
-func decodeJSON(r io.Reader, v interface{}) error {
-	defer io.Copy(io.Discard, r) //nolint:errcheck
-	d := json.NewDecoder(r)
-	// specify JSON numbers should get parsed to json.Number instead of float64 by default.
-	// This prevents loss between floats/ints.
-	d.UseNumber()
-	return d.Decode(v)
 }
