+++ internal/server/mcp_test.go
@@ -29,7 +29,7 @@ import (
 	"testing"
 
 	"github.com/googleapis/genai-toolbox/internal/log"
-	"github.com/googleapis/genai-toolbox/internal/server/mcp"
+	"github.com/googleapis/genai-toolbox/internal/server/mcp/jsonrpc"
 	"github.com/googleapis/genai-toolbox/internal/telemetry"
 )
 
@@ -76,16 +76,16 @@ func TestMcpEndpoint(t *testing.T) {
 		name  string
 		url   string
 		isErr bool
-		body  mcp.JSONRPCRequest
+		body  jsonrpc.JSONRPCRequest
 		want  map[string]any
 	}{
 		{
 			name: "initialize",
 			url:  "/",
-			body: mcp.JSONRPCRequest{
+			body: jsonrpc.JSONRPCRequest{
 				Jsonrpc: jsonrpcVersion,
 				Id:      "mcp-initialize",
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "initialize",
 				},
 			},
@@ -104,9 +104,9 @@ func TestMcpEndpoint(t *testing.T) {
 		{
 			name: "basic notification",
 			url:  "/",
-			body: mcp.JSONRPCRequest{
+			body: jsonrpc.JSONRPCRequest{
 				Jsonrpc: jsonrpcVersion,
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "notification",
 				},
 			},
@@ -114,10 +114,10 @@ func TestMcpEndpoint(t *testing.T) {
 		{
 			name: "tools/list",
 			url:  "/",
-			body: mcp.JSONRPCRequest{
+			body: jsonrpc.JSONRPCRequest{
 				Jsonrpc: jsonrpcVersion,
 				Id:      "tools-list",
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "tools/list",
 				},
 			},
@@ -146,10 +146,10 @@ func TestMcpEndpoint(t *testing.T) {
 		{
 			name: "tools/list on tool1_only",
 			url:  "/tool1_only",
-			body: mcp.JSONRPCRequest{
+			body: jsonrpc.JSONRPCRequest{
 				Jsonrpc: jsonrpcVersion,
 				Id:      "tools-list-tool1",
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "tools/list",
 				},
 			},
@@ -170,10 +170,10 @@ func TestMcpEndpoint(t *testing.T) {
 			name:  "tools/list on invalid tool set",
 			url:   "/foo",
 			isErr: true,
-			body: mcp.JSONRPCRequest{
+			body: jsonrpc.JSONRPCRequest{
 				Jsonrpc: jsonrpcVersion,
 				Id:      "tools-list-invalid-toolset",
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "tools/list",
 				},
 			},
@@ -190,10 +190,10 @@ func TestMcpEndpoint(t *testing.T) {
 			name:  "missing method",
 			url:   "/",
 			isErr: true,
-			body: mcp.JSONRPCRequest{
+			body: jsonrpc.JSONRPCRequest{
 				Jsonrpc: jsonrpcVersion,
 				Id:      "missing-method",
-				Request: mcp.Request{},
+				Request: jsonrpc.Request{},
 			},
 			want: map[string]any{
 				"jsonrpc": "2.0",
@@ -208,10 +208,10 @@ func TestMcpEndpoint(t *testing.T) {
 			name:  "invalid method",
 			url:   "/",
 			isErr: true,
-			body: mcp.JSONRPCRequest{
+			body: jsonrpc.JSONRPCRequest{
 				Jsonrpc: jsonrpcVersion,
 				Id:      "invalid-method",
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "foo",
 				},
 			},
@@ -228,10 +228,10 @@ func TestMcpEndpoint(t *testing.T) {
 			name:  "invalid jsonrpc version",
 			url:   "/",
 			isErr: true,
-			body: mcp.JSONRPCRequest{
+			body: jsonrpc.JSONRPCRequest{
 				Jsonrpc: "1.0",
 				Id:      "invalid-jsonrpc-version",
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "foo",
 				},
 			},
