+++ tests/alloydbainl/alloydb_ai_nl_integration_test.go
@@ -27,7 +27,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/googleapis/genai-toolbox/internal/server/mcp"
+	"github.com/googleapis/genai-toolbox/internal/server/mcp/jsonrpc"
 	"github.com/googleapis/genai-toolbox/tests"
 )
 
@@ -334,7 +334,7 @@ func runAiNlMCPToolCallMethod(t *testing.T) {
 	invokeTcs := []struct {
 		name          string
 		api           string
-		requestBody   mcp.JSONRPCRequest
+		requestBody   jsonrpc.JSONRPCRequest
 		requestHeader map[string]string
 		want          string
 	}{
@@ -342,10 +342,10 @@ func runAiNlMCPToolCallMethod(t *testing.T) {
 			name:          "MCP Invoke my-simple-tool",
 			api:           "http://127.0.0.1:5000/mcp",
 			requestHeader: map[string]string{},
-			requestBody: mcp.JSONRPCRequest{
+			requestBody: jsonrpc.JSONRPCRequest{
 				Jsonrpc: "2.0",
 				Id:      "my-simple-tool",
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "tools/call",
 				},
 				Params: map[string]any{
@@ -361,10 +361,10 @@ func runAiNlMCPToolCallMethod(t *testing.T) {
 			name:          "MCP Invoke invalid tool",
 			api:           "http://127.0.0.1:5000/mcp",
 			requestHeader: map[string]string{},
-			requestBody: mcp.JSONRPCRequest{
+			requestBody: jsonrpc.JSONRPCRequest{
 				Jsonrpc: "2.0",
 				Id:      "invalid-tool",
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "tools/call",
 				},
 				Params: map[string]any{
@@ -378,10 +378,10 @@ func runAiNlMCPToolCallMethod(t *testing.T) {
 			name:          "MCP Invoke my-auth-tool without parameters",
 			api:           "http://127.0.0.1:5000/mcp",
 			requestHeader: map[string]string{},
-			requestBody: mcp.JSONRPCRequest{
+			requestBody: jsonrpc.JSONRPCRequest{
 				Jsonrpc: "2.0",
 				Id:      "invoke-without-parameter",
-				Request: mcp.Request{
+				Request: jsonrpc.Request{
 					Method: "tools/call",
 				},
 				Params: map[string]any{
